vuln_detector/requirements.txt
streamlit==1.32.2
pandas==2.2.1
openai==1.14.2
anthropic==0.21.3
google-genai==0.3.0
python-dotenv==1.0.1
